Add files via upload
Login Php project I built on a LAMP stack with MariaDB on my dell latitude originally, moved over to my windows laptop with the PiZeroFileShare and the Login and Register Page are displaying correctly but I am having issues with the other pages, I think this may be due to the database its referencing does not exist on this computer, though I was having some issues before that. Still needing to fix but wanted to get the files up.